refactor(auth): JwtTokenProvider to JwtTokenService
- inject environment variable values at constructor
